Adding feature: rewrite of playlist in case of missing file or deletion
If files in the playlist do not exist anymore (and we tried to read them),
the program will fix the playlist and write an alternative playlist next to
the current one.
If the program is required to delete files, it will rewrite a new playlist
without the deleted files.

Those operations try to help to keep an up-to-date playlist.